modules/api/KafkaImpl/LocationProducer/requirements.txt
grpcio == 1.40.0
grpcio-tools == 1.40.0
kafka-python == 2.0.2